Move `getFragmentVisit` to plugin class
Avoids `getFragmentVisit.bind(this)`, which returns `any`
