+++ src/inc/handlers.ts
@@ -10,7 +10,6 @@ import {
 	getFirstMatchingRule,
 	cacheForeignFragmentElements,
 	shouldSkipAnimation,
-	getFragmentVisit,
 	adjustVisitScroll
 } from './functions.js';
 
@@ -36,7 +35,7 @@ export const onVisitStart: Handler<'visit:start'> = async function (this: Fragme
 	const route = getRoute(visit);
 	if (!route) return;
 
-	const fragmentVisit = getFragmentVisit.call(this, route);
+	const fragmentVisit = this.getFragmentVisit(route);
 
 	/**
 	 * Bail early if the current route doesn't match
